Fixes #23
The error reporting needs to be improved in the next iteration of
ProofScript (hopefully this will be a compiler then). In particular,
you not only want the location of where the error happened, but also
some sort of “stack trace” (or rather, function invocation trace)
leading up to the error.