Added ProofScriptVersion value to root theory.
The idea is that by declaring the version of ProofScript used in the
root theory file, we make ProofScript versioning a natural part of our
general versioning mechanism. We will need some special addition to the
ProofScript interpreter / compiler which detects this value and then
switches to the appropriate version of ProofScript.